+++ upcloud/error.go
@@ -47,7 +47,7 @@ func (err *Error) IsProblem() bool {
 }
 
 // Problem returns additional information on the problem with HTTP request (conforming to RFC7807), assuming they were included in error response.
-// Second return value is a boolean that will be true if the additional information actually exist.
+// Second return value is a boolean that will be true if the additional information actually exists.
 func (err *Error) Problem() (*ProblemError, bool) {
 	if err.IsProblem() {
 		return err.problemError, true
@@ -79,14 +79,14 @@ func (err *Error) UnmarshalJSON(buf []byte) error {
 		return unmarshalErr
 	}
 
-	// We got the legacy error struct from the API, just populate the main error fields
+	// We got the legacy error struct from the API, just populate the main error fields and return
 	if localError.LegacyError != nil {
 		err.Code = localError.LegacyError.ErrorCode
 		err.Message = localError.LegacyError.ErrorMessage
 		return nil
 	}
 
-	// This means we got a json+problem error from the API, populate the fields accordingly
+	// We got a json+problem error from the API, populate the problem fields accordingly
 	err.problemError = &ProblemError{
 		Status:        localError.ProblemStatus,
 		Title:         localError.ProblemTitle,
@@ -102,23 +102,6 @@ func (err *Error) UnmarshalJSON(buf []byte) error {
 	return nil
 }
 
-//// UnmarshalJSON is a custom unmarshaller that deals with
-//// deeply embedded values.
-//func (e *Error) UnmarshalJSON(b []byte) error {
-//	type localError Error
-//	v := struct {
-//		Error localError `json:"error"`
-//	}{}
-//	err := json.Unmarshal(b, &v)
-//	if err != nil {
-//		return err
-//	}
-//
-//	(*e) = Error(v.Error)
-//
-//	return nil
-//}
-
 // ProblemError is the type conforming to RFC7807 that represents an error or a problem associated with an HTTP request.
 type ProblemError struct {
 	// Type is the URI to a page describing the problem
@@ -128,10 +111,16 @@ type ProblemError struct {
 	// InvalidParams if set, is a list of ProblemInvalidParam describing a specific part(s) of the request
 	// that caused the problem
 	InvalidParams []ProblemInvalidParam `json:"invalid_params,omitempty"`
-	// CorrelationID is an unique string that identifies the request that caused the problem
+	// CorrelationID is a unique string that identifies the request that caused the problem
 	CorrelationID string `json:"correlation_id,omitempty"`
 	// HTTP Status code
 	Status int `json:"status"`
+}
+
+// ProblemErrorInvalidParam is a type describing extra information in the Problem type's InvalidParams field.
+type ProblemErrorInvalidParam struct {
+	Name   string `json:"name"`
+	Reason string `json:"reason"`
 }
 
 // Error implements the Error interface
@@ -157,10 +146,4 @@ func (pe *ProblemError) getShortenedType() string {
 	}
 
 	return strings.Replace(parts[1], "ERROR_", "", 1)
-}
-
-// ProblemErrorInvalidParam is a type describing extra information in the Problem type's InvalidParams field.
-type ProblemErrorInvalidParam struct {
-	Name   string `json:"name"`
-	Reason string `json:"reason"`
 }