+++ upcloud/request/kubernetes.go
@@ -48,11 +48,11 @@ func (r *GetKubernetesClusterRequest) RequestURL() string {
 
 // CreateKubernetesClusterRequest represents a request to create a Kubernetes cluster
 type CreateKubernetesClusterRequest struct {
-	Name        string                        `json:"name"`
-	Network     string                        `json:"network"`
-	NetworkCIDR string                        `json:"network_cidr"`
-	NodeGroups  []upcloud.KubernetesNodeGroup `json:"node_groups"`
-	Zone        string                        `json:"zone"`
+	Name        string                `json:"name"`
+	Network     string                `json:"network"`
+	NetworkCIDR string                `json:"network_cidr"`
+	NodeGroups  []KubernetesNodeGroup `json:"node_groups"`
+	Zone        string                `json:"zone"`
 }
 
 func (r *CreateKubernetesClusterRequest) RequestURL() string {